% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchAlign5.R
\name{aggregateIndex}
\alias{aggregateIndex}
\title{BA: Aggregate clustering information from two sample types}
\usage{
aggregateIndex(aI1, aI2)
}
\arguments{
\item{aI1}{an alignIndex object for sample type 1}

\item{aI2}{an alignIndex object for sample type 2}
}
\value{
An object (list) with aggregated alignment information
}
\description{
aggregateIndex will aggregate alignment indices IF present for both sample types
}

