% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchAlign5.R
\name{align}
\alias{align}
\title{Find alignment candidates}
\usage{
align(flags, mz, rt, mzdiff = 0.005, rtdiff = 10)
}
\arguments{
\item{flags}{a table of presentness/missingness per batch x sample type}

\item{mz}{m/z values of the features}

\item{rt}{rt values of the features}

\item{mzdiff}{maximum distance in m/z to be considered for alignment}

\item{rtdiff}{maximum distance in rt to be considered for alignment}
}
\value{
events: all possible alignments

features: all features involved in possible alignments

clusters: groups of features linked through mutual events
}
\description{
align will find candidates for alignment among features that have systematic missingness "within box" of maximum m/z & rt differences. Used internally.
}

