% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchAlign5.R
\name{alignIndex}
\alias{alignIndex}
\title{BA: Find alignment candidates}
\usage{
alignIndex(batchflag, flagType = c("Hard", "Soft", "All"), grpType = "Q",
  mzdiff = 0.005, rtdiff = 10, report = TRUE, reportName = "splits")
}
\arguments{
\item{batchflag}{a table of presentness/missingness per batch x sample type for the features within cluster}

\item{flagType}{This is experimental. Use at your own risk.}

\item{grpType}{sample type to be used to find batch alignments}

\item{mzdiff}{maximum distance in m/z to be considered for alignment}

\item{rtdiff}{maximum distance in rt to be considered for alignment}

\item{report}{boolean for plotting results of owercrowded cluster splitting for visual examination}

\item{reportName}{string with name of report file}
}
\value{
An object (list) consisting of the following features:

grpType: sample type used to find batch alignments

shift: list with features to combine

events: all possible alignment candidates (events)

features: features after splitting overcrowded clusters

clusters: clusters after splitting overcrowded clusters

oldFeatures: features before splitting overcrowded clusters

oldClusters: clusters before splitting overcrowded clusters
}
\description{
alignIndex will find features systematically misaligned between batches using "sample type" information.
}
