% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchAlign5.R
\name{batchAlign}
\alias{batchAlign}
\title{BA: Alignment of peaktable based on alignIndex and batchFlag data}
\usage{
batchAlign(batchflag, alignindex, peaktable_filled, meta)
}
\arguments{
\item{batchflag}{a table of presentness/missingness per batch x sample type for the features within cluster}

\item{alignindex}{An object (list) consisting of alignment information}

\item{peaktable_filled}{a peaktable without missing values (ie after hard filling or imputation)}

\item{meta}{a matrix of metadata with batch in column 1 and sample type in column 2}
}
\value{
An object (list) consisting of the following features:

PTalign: A peaktable with batch aligned peak areas

boolAveragedAlign: boolean vector of features where alignment has been made using feature averaging (i.e. where batches are missing within features). Length: same as final number of features

PTfill: peaktable without missing values (indata)

boolKeep: boolean vector of features kept after alignment (the rest of the combined features are deleted)

boolAveragedFill: boolean vector of features where alignment has been made using feature averaging (i.e. where batches are missing within features). Length: same as orgiginal number of features

aI: alignIndex object (indata)

meta: metadata (indata)
}
\description{
batchAlign will use the 'batchFlag' and 'alignIndex' information to align sample peaks that are systematically misaligned across batches.
}

