% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchNorm1.r
\name{batchComb}
\alias{batchComb}
\title{Extract features from multiple batch data}
\usage{
batchComb(batchObjs, batchLimit, finalFeats)
}
\arguments{
\item{batchObjs}{a list with batch objects (from within batch drift correction)}

\item{batchLimit}{lower limit of number of batches in which the feature needs to be present to be included in the final list (defaults to length of `batchFeats` list)}

\item{finalVars}{a vector with the features to bring out from batch PTs. NB! All features need be present in all batches (defaults to all features present in all batches)}
}
\value{
a combined, but NOT normalised, peaktable - limited to common features
}
\description{
Extract features present in all batches and combine them into a master peaktable.
}

