% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchAlign5.R
\name{batchFlag}
\alias{batchFlag}
\title{Aggregate presense/missingness per batch}
\usage{
batchFlag(PTnofill, meta, NAhard = 0.8, NAsoft = 0.5, quantileSoft = 0.1)
}
\arguments{
\item{PTnofill}{a peaktable with missing values (ie without either hard filling or imputation)}

\item{meta}{a matrix of metadata with batch in column 1 and sample type in column 2}

\item{NAhard}{proportion of NAs within batch for feature to be considered missing}

\item{NAsoft}{experimental: Use at own risk}

\item{quantileSoft}{experimental: Use at own risk}
}
\value{
meta: metadata aggretated by batch x sample type

flagHard: a table of presentness/missingness per batch x sample type

mz: m/z values of the features

rt: rt values of the features

flagAll: experimental: Use at own risk

flagSoft: experimental: Use at own risk

flagType: experimental: Use at own risk
}
\description{
batchFlag will aggregate presentness/missingness per feature at batch level within the combination: batch x sample type
}

