% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustFuncBare3.r
\name{cleanVar}
\alias{cleanVar}
\title{DC: Remove features not passing QC test}
\usage{
cleanVar(QCCorr, CVlimit = 0.2, report = FALSE)
}
\arguments{
\item{QCCorr}{a Corr object}

\item{CVlimit}{user-defined QC riterion to pass. Defaults to 0.2}

\item{report}{boolean whether to print a pdf report of drift models}
}
\value{
A Clean object consisting of:

original information from Corr object (indata) and

updated actionInfo (actions taken per cluster)

QCFeatsFinal: Final peaktable (scaled) for QC samples

TestFeatsFinal: Final peaktable for batch samples

finalVars: All features kept within the final peaktable

QCcvs: CVs per features kept within the final peaktable
}
\description{
Remove those features with CV(QC)>limit (defined by user)
}
