% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustFuncBare3.r
\name{clust}
\alias{clust}
\title{DC: Cluster features}
\usage{
clust(QCInjs, QCFeats, modelNames = c("VVE"), G = seq(1, 52, by = 3),
  report = FALSE)
}
\arguments{
\item{QCInjs}{vectors with QC injections in sequence}

\item{QCFeats}{Feature matrix for QC injections: Scaled (but not centered) features as columns; Injections as rows}

\item{modelNames}{Which 'mclust' models to consider (see mclust package for details)}

\item{G}{Which number of clusters to consider (see mclust package for details)}

\item{report}{boolean whether to print a pdf report of clustering results}
}
\value{
a clust object containing:

QCInjs: as indata

QCFeats: as indata

BIC: Bayesian Information Criteria-object for the investigated clustering models

clust: summary of the BIC-object (i.e. clustering)

BICTime: Time required for the BIC-calculation

clustTime: Time required for the clustering
}
\description{
clust will perform clustering of features with similar drift pattern by projecting scaled QC features as coordinates in observation/injection space
}

