% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchAlign5.R
\name{clusterMatrix}
\alias{clusterMatrix}
\title{BA: Sort features (alignment candidates) into clusters}
\usage{
clusterMatrix(features)
}
\arguments{
\item{features}{all features involved in possible alignments}
}
\value{
clusters: groups of features linked through mutual events
}
\description{
clusterMatrix will sort features into clusters. Used internally.
}

