% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchAlign5.R
\name{clustSplit}
\alias{clustSplit}
\title{Split overcrowded clusters into subclusters}
\usage{
clustSplit(clustFlags, mz, rt)
}
\arguments{
\item{clustFlags}{a table of presentness/missingness per batch x sample type for the features within cluster}

\item{mz}{m/z values of the features}

\item{rt}{rt values of the features}
}
\value{
A vector with subcluster identifiers
}
\description{
clustSplit will split clusters by iteratively removing maximum distances until "single possible candidates" are identified. The function works recursively of severely overcrowded clusters. Used internally.
}

