% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctDrift.R
\name{correctDrift}
\alias{correctDrift}
\title{Within-batch signal intensity drift correction}
\usage{
correctDrift(
  peakTable,
  injections,
  sampleGroups,
  QCID = "QC",
  RefID = "none",
  modelNames = c("VVV", "VVE", "VEV", "VEE", "VEI", "VVI", "VII"),
  G = seq(5, 35, by = 10),
  smoothFunc = "spline",
  spar = 0.2,
  CVlimit = 0.3,
  report = TRUE
)
}
\arguments{
\item{peakTable}{Batch peak table}

\item{injections}{Injection number in sequence}

\item{sampleGroups}{Vector (length=nrow(PeakTab)) of sample type (e.g. "sample", "QC", "Ref)}

\item{QCID}{QC identifier in sampleGroups (defaults to "QC")}

\item{RefID}{Optional identifier in sampleGroups of external reference samples for unbiased assessment of drift correction (see batchCorr tutorial at Gitlab page)}

\item{modelNames}{Which MClust geometries to test (see mclust documentation)}

\item{G}{Which numbers of clusters to test (see mclust documentation)}

\item{smoothFunc}{choice of regression function: spline or loess (defaults to spline)}

\item{spar}{smoothing parameter for spline or loess (defaults to 0.2)}

\item{CVlimit}{QC feature CV limit as final feature inclusion criterion}

\item{report}{Boolean whether to print pdf reports of drift models}
}
\value{
A driftCorrection object

$actionInfo (to see what happened to each cluster)

$testFeatsCorr (to extract drift-corrected data) and

$testFeatsFinal (to extract drift-corrected data which pass the criterion that QC CV < CVlimit.
}
\description{
Within-batch signal intensity drift correction
}
\examples{
data('ThreeBatchData') 
set.seed(2024)
# Get batches
batchB <- getBatch(peakTable = PTfill, meta = meta, 
                   batch = meta$batch, select = 'B')
batchF <- getBatch(peakTable = PTfill, meta = meta, 
                   batch = meta$batch, select = 'F')
# Drift correction using QCs
BCorr <- correctDrift(peakTable = batchB$peakTable, 
                      injections = batchB$meta$inj, 
                      sampleGroups = batchB$meta$grp, QCID = 'QC', 
                      G = seq(5,35,by=3), modelNames = c('VVE', 'VEE'))
# More unbiased drift correction using QCs & external reference samples
FCorr <- correctDrift(peakTable = batchF$peakTable, 
                      injections = batchF$meta$inj,
                      sampleGroups = batchF$meta$grp, QCID = 'QC',
                      RefID='Ref', G = seq(5,35,by=3), 
                      modelNames = c('VVE', 'VEE'))
# Merge batches for batch normalization, for example
mergedData <- mergeBatches(list(BCorr, FCorr))
}
