% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustFuncBare3.r
\name{cv}
\alias{cv}
\title{Coefficient of variation (CV)}
\usage{
cv(mat)
}
\arguments{
\item{mat}{a matrix with variables as columns}
}
\value{
a vector of CVs
}
\description{
Calculates CV per column in a matrix. Aka relative standard deviation (RSD).
}

