% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustFuncBare3.r
\name{driftCalc}
\alias{driftCalc}
\title{DC: Calculate intensity drift per cluster}
\usage{
driftCalc(
  QCClust,
  smoothFunc = c("spline", "loess"),
  spar = 0.2,
  report = FALSE
)
}
\arguments{
\item{QCClust}{a clust object}

\item{smoothFunc}{choice of regression function: spline or loess (defaults to spline)}

\item{spar}{smoothing parameter for spline or loess}

\item{report}{boolean whether to print a pdf report of drift models}
}
\value{
a driftCalc object containing:

original information from clust object (indata) and

actionInfo

ratios

corMat

deltaDist

varClust
}
\description{
Clustered, scaled QC features are pooled together and intensity drift patterns are modelled per cluster
}
