% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustFuncBare3.r
\name{driftCorr}
\alias{driftCorr}
\title{DC: Correct for intensity drift per cluster}
\usage{
driftCorr(QCDriftCalc, refList = NA, refType = c("none", "one", "many"),
  CorrObj = NA, report = FALSE)
}
\arguments{
\item{QCDriftCalc}{a DriftCalc object}

\item{refList}{a reference sample onject with 'inj' and 'Feats'}

\item{refType}{at present, the options "one" and "none" are supported for one or no reference samples present.}

\item{CorrObj}{a batch object to be corrected with 'inj' and 'Feats'. If not present defaults to the QC object}

\item{report}{boolean whether to print a pdf report of drift models}
}
\value{
A Corr object consisting of:

original information from DriftCalc object (indata) and

updated actionInfo (actions taken per cluster)

QCFeatsCorr Drift-corrected QC features

refType (indata)

drift corrected reference samples

drift corrected corr/batch samples
}
\description{
Perform signal intensity drift correction if resulting in increased quality of data (measured by reduced root-mean-squared distances of reference samples).
}
