% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustFuncBare3.r
\name{driftWrap}
\alias{driftWrap}
\title{DC: Perform drift correction}
\usage{
driftWrap(QCObject, modelNames = NULL, G = seq(1, 40, by = 3), BatchObject,
  RefObject = NULL, CVlimit = 0.3, report = FALSE)
}
\arguments{
\item{QCObject}{QC Object (as obtained from `makeQCObject()`)}

\item{modelNames}{Which MClust geometries to test (see mclust documentation)}

\item{G}{Which numbers of clusters to test (see mclust documentation)}

\item{BatchObject}{Batch object (as obtained from `makeBatchObject()`) to be corrected for drift}

\item{RefObject}{Optional reference object (as obtained from `makeBatchObject()`) to validate whether QC correction improves data quality}

\item{CVlimit}{QC feature CV limit as final feature inclusion criterion}

\item{report}{boolean whether to print a pdf report of drift models}
}
\value{
A drift corrected object with QC features CV<limit containing final peak table
}
\description{
Wrapper function for all drift subfunctions (clust, driftCalc, driftCorr, cleanVar)
}
