% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustFuncBare3.r
\name{driftWrap3}
\alias{driftWrap3}
\title{DC: Perform drift correction}
\usage{
driftWrap3(grabObj, refType, CVlimit = 0.3, report = FALSE)
}
\arguments{
\item{grabObj}{a grabWrap object}

\item{refType}{type of reference sample strategy used (at present "one" or "none" supported)}

\item{CVlimit}{QC feature CV limit as final feature inclusion criterion}

\item{report}{boolean whether to print a pdf report of drift models}
}
\value{
A drift corrected object with QC features CV<limit containing final peak table
}
\description{
Wrapper function for all drift subfunctions (clust, driftCalc, driftCorr, cleanVar)
}

