% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchNorm1.r
\name{featComb}
\alias{featComb}
\title{Limit variables/features to those common between batches}
\usage{
featComb(batchFeats, batchLimit)
}
\arguments{
\item{batchFeats}{list with feature names per batch}

\item{batchLimit}{lower limit of number of batches in which the feature needs to be present to be included in the final list (defaults to length of `batchFeats` list)}
}
\value{
a vector with the features present in at least `batchLimit` batches
}
\description{
Extract features from peak table and report their m/z and rt values
}

