% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gets.R
\name{getGroup}
\alias{getGroup}
\title{Extract specific sample group from peaktable and metadata}
\usage{
getGroup(peakTable, meta, sampleGroup, select)
}
\arguments{
\item{peakTable}{Multi-batch peak table}

\item{meta}{Multi-batch metadata (including e.g. batch, injection sequence, sample tye, sample name, ...)}

\item{sampleGroup}{vector (length = nSamples) containing sample group information (e.g. QC, Sample, Reference)}

\item{select}{which sample group to extract}
}
\value{
list object

$peakTable: Sample group peak table

$meta: Sample group metadata
}
\description{
Matches pattern in `select` to sample group and extracts
}
\examples{
data(ThreeBatchData)
batchB=getBatch(peakTable=PTfill, meta=meta, batch=meta$batch, select='B')
batchBQC=getGroup(peakTable=batchB$peakTable, meta=batchB$meta, sampleGroup=batchB$Meta$grp, select='QC')
}
