% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustFuncBare3.r
\name{grabRef}
\alias{grabRef}
\title{DC: Grab Reference samples}
\usage{
grabRef(XS, QC, grp = "Ref")
}
\arguments{
\item{XS}{An XCMS object}

\item{QC}{a grabQC object}

\item{grp}{an identifyer for the reference samples}
}
\value{
a list containing:

inj: a vector of reference sample injection numbers in sequence

Feats: Features (peak table)
}
\description{
Bring out reference sample group from an XCMS object: Exclude features missing from QC samples
}

