% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustFuncBare3.r
\name{grabWrap}
\alias{grabWrap}
\title{DC: Grab samples for drift correction}
\usage{
grabWrap(XS, batch, QC = "QC", Ref = "Ref")
}
\arguments{
\item{XS}{an XCMS object}

\item{batch}{a batch identifier}

\item{QC}{a QC sample identifier}

\item{Ref}{a reference sample identifier}
}
\value{
An object containing:

QC: A QC object

Ref: A Ref object

Batch: A batch object
}
\description{
Wrapper function for grabbing QCs, reference and entire batch samples from XCMS-set
}

