% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustFuncBare3.r
\name{grabWrapBA}
\alias{grabWrapBA}
\title{DC: Grab samples for drift correction after batch alignment}
\usage{
grabWrapBA(XS, PT, batch, QC = "QC", Ref = "Ref")
}
\arguments{
\item{XS}{an XCMS object}

\item{PT}{a master peaktable after batch alignment}

\item{batch}{a batch identifier}

\item{QC}{a QC sample identifier}

\item{Ref}{a reference sample identifier}
}
\value{
An object containing:

QC: A QC object

Ref: A Ref object

Batch: A batch object
}
\description{
Wrapper function for grabbing QCs, reference and entire batch samples combining information from XCMS object and batch-aligned peaktable.
}
