% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeBatchObject.R
\name{makeBatchObject}
\alias{makeBatchObject}
\title{Prepare a Batch object for drift correction}
\usage{
makeBatchObject(peakTable, inj, QCObject)
}
\arguments{
\item{peakTable}{Peak table (without missing values). Samples in rows; Features in columns; colnames as mz@rt}

\item{inj}{Injection sequence number for Batch samples}

\item{QCObject}{QC Object (to check injection sequence compatibility)}
}
\value{
Batch Object
}
\description{
Prepare a Batch object for drift correction
}
\examples{
BatchObject=makeBatchObject(BatchPeakTable, BatchInjections, QCObject)
}
