% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeQCObject.R
\name{makeQCObject}
\alias{makeQCObject}
\title{Prepare a QC object for drift correction}
\usage{
makeQCObject(QCPeakTab, inj)
}
\arguments{
\item{QCPeakTab}{Peak table (without missing values). Samples in rows; Features in columns; colnames as mz@rt}

\item{inj}{Injection sequence number for QC samples}
}
\value{
QC object
}
\description{
Prepare a QC object for drift correction
}
\examples{
QCObject=makeQCObject(QCPeakTable, QCInjections)
}
