% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeBatches.R
\name{mergeBatches}
\alias{mergeBatches}
\title{Merge batches after drift correction}
\usage{
mergeBatches(batchList, qualRatio = 0.5)
}
\arguments{
\item{batchList}{A list of correction objects (after drift correction)}

\item{qualRatio}{Proportion of}
}
\value{
A list object containing

`peakTableOrg` A merged peak table of original data

`peakTableCorr` A merged peak table of drift-corrected data

`batch` Batch identifier (per sample)

`injection` Injection number (per sample)
}
\description{
The output of the within-batch drift correction is a correction object.
This function merges peak tables from several batches by extracting information from the correction objects.
The user must specify a minimum proportion of qualified batches per feature, i.e. such batches where the QC CV is < the specified limit.
There is thus a risk that features with poor quality (in certain batches) are present, but the features are present in high quality in sufficient proportion of batches to anyway warrant inclusion.
}
\examples{
batchList <- list(b1=b1Corr, b2=b2Corr, b3=b3Corr)
mergeList <- mergeBatches(batchList)
cbind(mergeList$batch,mergeList$injection)
dim(mergeList$peakTable)
}
