% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{meta}
\alias{meta}
\title{meta}
\format{A data frame with 90 observations of 2 variables
  (i.e. metabolite features) - containing missing values:
  \code{batch}, \code{grp}.}
\usage{
meta
}
\description{
Metadata for the injections in PTfill and PTnofill
}
\details{
Col 1: Batch identifier

Col 2: Sample type identifier (Q for QC and R for Ref)
}
\keyword{datasets}

