% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizeBatches.R
\name{normalizeBatches}
\alias{normalizeBatches}
\title{Between-batch normalisation}
\usage{
normalizeBatches(peakTable, batches, sampleGroup, refGroup = "QC",
  population = "all", CVlimit = 0.3, FCLimit = 5)
}
\arguments{
\item{peakTable}{Multi-batch peak table (samples in rows; features in columns)}

\item{batches}{Vector (length=nrow(PeakTab)) of batch identifiers}

\item{sampleGroup}{Vector (length=nrow(PeakTab)) of sample type (e.g. "sample", "QC", "Ref)}

\item{refGroup}{Identifier of reference samples in sampleGroups (defaults to "QC")}

\item{population}{Identifier of population samples in sampleGroups (all (default) or any type of samples present in sampleGroups)}

\item{CVlimit}{CV criterion to pass for Ref samples per batch (defaults to 0.3)}

\item{FCLimit}{Fold-change criterion for intensity (in relation to average intensity FC between batches; defaults to 5)}
}
\value{
An list object containing:

peakTable: Normalised multi-batch peak table

refCorrected: Boolean matrix with info on which batches were normalised by reference samples; others were normalized by population medians
}
\description{
Batches are feature-wise normalised by Ref samples if passing heuristic criteria (CV and fold change).
Otherwise normalized by population median (still per feature).
}
