% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchAlign5.R
\name{peakTab}
\alias{peakTab}
\title{Get a peak table from an XCMS object}
\usage{
peakTab(XS)
}
\arguments{
\item{XS}{an XCMS object}
}
\value{
a peak table with samples as rows and features as columns, features (colnames) named by 'm/z@rt'
}
\description{
peakTab will extract a peak table from an XCMS object with samples as rows and features as columns. Features are named by 'm/z@rt'
}
