% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchAlign5.R
\name{plotAlign}
\alias{plotAlign}
\title{BA: Plot clusters of aligned features}
\usage{
plotAlign(batchflag, alignindex, clust, plotType = c("plot", "pdf",
  "png"), reportName = "clustPlot", mzwidth = 0.02, rtwidth = 100)
}
\arguments{
\item{batchflag}{a table of presentness/missingness per batch x sample type for the features within cluster}

\item{alignindex}{An object (list) consisting of alignment information}

\item{clust}{which cluster(s) to plot. If missing, plots all clusters.}

\item{plotType}{Whether to plot internally ('plot'), to pdf ('pdf') or png ('png') file.}

\item{reportName}{name of plotfile}

\item{mzwidth}{plot span of m/z}

\item{rtwidth}{plot span of rt}
}
\description{
plotAlign will plot clusters of aligned features
}
