% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchAlign5.R
\name{plotClust}
\alias{plotClust}
\title{BA: Plot the identified clusters}
\usage{
plotClust(batchflag, grpFlag, cluster, text, color = 2, mzwidth = 0.02,
  rtwidth = 100)
}
\arguments{
\item{grpFlag}{a vector corresponding to the rows of batchFlag used for sample type}

\item{cluster}{extracted cluster to be plotted}

\item{text}{cluster identifier}

\item{color}{vector with colors}

\item{mzwidth}{plot span of m/z}

\item{rtwidth}{plot span of rt}

\item{batchFlag}{a table of presentness/missingness per batch x sample type for the features within cluster}
}
\description{
plotClust will plot clusters identified based on sample type. Used internally.
}

