% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PTfill}
\alias{PTfill}
\title{PTfill}
\format{A data frame with 90 observations of 11667 variables
  (i.e. metabolite features) - without missing values}
\usage{
PTfill
}
\description{
Filled peaktable from XCMS object containing batch-specific QCs and
long-term reference samples from three batches of LC-MS metabolomics data.
Peaktable contains all values, i.e. after hard integration of EICs.
}
\keyword{datasets}

