% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PTnofill}
\alias{PTnofill}
\title{PTnofill}
\format{A data frame with 90 observations of 11667 variables
  (i.e. metabolite features) - containing missing values}
\usage{
PTnofill
}
\description{
Unfilled peaktable from XCMS object containing batch-specific QCs and
long-term reference samples from three batches of LC-MS metabolomics data.
Peaktable contains missing values, used for the batch alignment algorithm.
}
\keyword{datasets}

