% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchNorm1.r
\name{refOut}
\alias{refOut}
\title{BN: Info on reference samples aggregated on batch level (Deprecated)}
\usage{
refOut(PT, meta, grpType = "R", CVlimit = 0.3)
}
\arguments{
\item{PT}{a multi-batch master peak table}

\item{meta}{metadata with batch (col1) and sample type (col2)}

\item{grpType}{a sample type identifier for reference samples}

\item{CVlimit}{CV criterion to pass for Ref samples per batch}
}
\value{
an object containing:

CV: boolean per batch & feature in CV<limit

aveInt: average reference intensity per batch & feature
}
\description{
Reference samples are aggregated on batch level
}
