% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustFuncBare3.r
\name{rmsDist}
\alias{rmsDist}
\title{Root mean squared distance}
\usage{
rmsDist(mat)
}
\arguments{
\item{mat}{a matrix containing observations as rows and variables as columns}
}
\value{
a numeric rmsDist
}
\description{
Calculate root mean squared distance from center point
}

