% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchAlign5.R
\name{alignBatches}
\alias{alignBatches}
\title{Perform multi-batch alignment to merge features artificially split between batches}
\usage{
alignBatches(peakInfo, PeakTabNoFill, PeakTabFilled, batches, sampleGroups,
  selectGroup, mzdiff = 0.002, rtdiff = 15, report = TRUE)
}
\arguments{
\item{peakInfo}{matrix with mz and rt in columns 1:2 (see e.g. ?peakInfo)}

\item{PeakTabNoFill}{Multi-batch peak table including missing values}

\item{PeakTabFilled}{Multi-batch peak table without missing values}

\item{batches}{Vector (length=nrow(PeakTab)) of batch identifiers}

\item{sampleGroups}{Vector (length=nrow(PeakTab)) of sample type (e.g. "sample", "QC", "Ref)}

\item{selectGroup}{Which sample type to base alignment on (e.g. "sample", "QC" or "Ref")}

\item{mzdiff}{Tolerance for difference in m/z}

\item{rtdiff}{Tolerance for difference in retention time}

\item{report}{Whether to export diagnostic plots into your work directory (defaults to TRUE)}
}
\value{
batchAlign Object

Aligned peak table available under $PTalign

Returns NULL if no alignment candidates can be found
}
\description{
Perform multi-batch alignment to merge features artificially split between batches
}
\examples{
library(batchCorr)
data('ThreeBatchData') # loads "meta" (metadata), "PTnofill" (Multi-batch peak table including missing values) and "PTfill" (Multi-batch peak table without missing values)
peakIn=peakInfo(PT = PTnofill, sep = '@', start = 3) # Extract peakinfo (i.e. m/z and rt of features) # Thesecolumn names have 2 leading characters describing LC-MS mode -> start at 3
alignBat <- alignBatches(peakInfo = peakIn, PeakTabNoFill = PTnofill, PeakTabFilled = PTfill, batches = meta$batch, sampleGroups = meta$grp, selectGroup = 'QC') # Perform multi-batch alignment
PTAlign=alignBat$PTalign # Extract new peak table
}
