% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctDrift.R
\name{correctDrift}
\alias{correctDrift}
\title{Within-batch signal intensity drift correction}
\usage{
correctDrift(peakTable, injections, sampleGroups, QCID = "QC",
  RefID = "none", modelNames = c("VVV", "VVE", "VEV", "VEE", "VEI",
  "VVI", "VII"), G = seq(5, 35, by = 10), CVlimit = 0.3,
  report = TRUE)
}
\arguments{
\item{peakTable}{Batch peak table}

\item{injections}{Injection number in sequence}

\item{sampleGroups}{Vector (length=nrow(PeakTab)) of sample type (e.g. "sample", "QC", "Ref)}

\item{QCID}{QC identifier in sampleGroups (defaults to "QC")}

\item{RefID}{Optional identifier in sampleGroups of external reference samples for unbiased assessment of drift correction (see batchCorr tutorial at Gitlab page)}

\item{modelNames}{Which MClust geometries to test (see mclust documentation)}

\item{G}{Which numbers of clusters to test (see mclust documentation)}

\item{CVlimit}{QC feature CV limit as final feature inclusion criterion}

\item{report}{Boolean whether to print pdf reports of drift models}
}
\value{
A driftCorrection object

$actionInfo (to see what happened to each cluster)

$testFeatsCorr (to extract drift-corrected data) and

$testFeatsFinal (to extract drift-corrected data which pass the criterion that QC CV < CVlimit.
}
\description{
Within-batch signal intensity drift correction
}
\examples{
library(batchCorr)
data('OneBatchData') # loads "B_meta" (metadata), "B_PT" (Peak table without missing values)
# Drift correction using QCs -> Approx 2-3 minutes computation
batchBCorr <- correctDrift(peakTable = B_PT, injections = B_meta$inj, sampleGroups = B_meta$grp, QCID = 'QC', modelNames = 'VVE', G = 17:22)
# More unbiased drift correction using QCs & external reference samples -> Approx 2-3 minutes computation
batchBCorr <- correctDrift(peakTable = B_PT, injections = B_meta$inj, sampleGroups = B_meta$grp, QCID = 'QC', RefID='Ref', modelNames = 'VVE', G = 17:22)
}
