% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustFuncBare3.r
\name{grabBatch}
\alias{grabBatch}
\title{DC: Grab Batch samples}
\usage{
grabBatch(XS, QC)
}
\arguments{
\item{XS}{An XCMS object}

\item{QC}{a grabQC object}
}
\value{
a list containing:

inj: a vector of sample injection numbers in sequence

Feats: Features (peak table)
}
\description{
Bring out all batch samples from an XCMS object: Exclude features missing from QC samples
}

