% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustFuncBare3.r
\name{grabQC}
\alias{grabQC}
\title{DC: Grab QC samples}
\usage{
grabQC(XS, batch, grp = "QC")
}
\arguments{
\item{XS}{An XCMS object}

\item{batch}{a batch identifier}

\item{grp}{an identifyer for the QC samples}
}
\value{
a list containing:

batch: batch (as input)

inj: a vector of QC injection numbers in sequence

Feats: Scaled features (peak table)

RawFeats: Unscaled features including 100% NA features (peak table)

RawFeatsNaRm: Unscaled features excluding 100% NA features (peak table)

NAs: which features are missing from QC samples
}
\description{
Bring out 'QC' group from an XCMS object: Scaled and 100% NAs removed
}

