% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchNorm1.r
\name{peakInfo}
\alias{peakInfo}
\title{Extract m/z and rt from peak table}
\usage{
peakInfo(PT)
}
\arguments{
\item{PT}{a peak table with variables as columns}
}
\value{
a matrix with m/z and rt of features as columns
}
\description{
Extract features from peak table and report their m/z and rt values
}

