% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchNorm1.r
\name{refCorr}
\alias{refCorr}
\title{BN: Between batch normalisation by Ref samples}
\usage{
refCorr(PT, meta, refs, FCLimit = 5)
}
\arguments{
\item{PT}{a multi-batch master peak table}

\item{meta}{metadata with batch (col1) and sample type (col2)}

\item{refs}{a refOut object}

\item{FClimit}{Fold-change criterion for intensity (in relation to average intensity FC between batches)}
}
\value{
an object containing:

PTRef: Reference sample-normalised multi-batch peak table

refCorr: Boolean matrix with info on which batches were normalised by reference samples

PTOrg (indata peaktable)
}
\description{
Batches are normalised by Ref samples if passing heuristic distance criterion.
}

